Command mv changed: moving multi files is supported

+++ bin/mv.lua
@@ -1,25 +1,43 @@
 local fs = require("filesystem")
 local shell = require("shell")
+local buffer = require("buffer")
+local messages = require("messages")
 
 local args, options = shell.parse(...)
 if #args < 2 then
-  io.write("Usage: mv [-f] <from> <to>\n")
-  io.write(" -f: overwrite file if it already exists.")
-  return
+	io.write("Usage: mv [-f] <from> [...] <to>\n")
+	io.write("	-f: Overwrite file if it already exists.")
+	return
 end
 
-local from = shell.resolve(args[1])
-local to = shell.resolve(args[2])
-if fs.isDirectory(to) then
-  to = to .. "/" .. fs.name(from)
+local to = shell.resolve(args[#args])
+local todir = fs.isDirectory(to)
+
+if not todir and #args > 2 then
+	buffer.write(io.stderr, "mv: " .. messages.ENOTDIR)
+	return
 end
-if fs.exists(to) then
-  if not options.f then
-    io.stderr:write("target file exists")
-  end
-  fs.remove(to)
-end
-local result, reason = os.rename(from, to)
-if not result then
-  io.stderr:write(reason or "unknown error")
+
+local i = 1
+while i < #args do
+	local from = shell.resolve(args[i])
+	if todir then
+		to = to .. "/" .. fs.name(from)
+	end
+	if fs.exists(to) then
+		if not options.f then
+			buffer.write(io.stderr, "target file exists")
+			return
+		end
+		fs.remove(to)
+	end
+
+	local result, reason = os.rename(from, to)
+	if not result then
+		buffer.write(io.stderr, reason or "unknown error")
+		if not options.f then
+			return
+		end
+	end
+	i = i + 1
 end
